Content Updates
Content folder started.

I added classes under the typography folder for font sizes.
I think this makes sense to do it this way but I could be wrong.  Being able to add a class to an element to define it's font size means we don't have to constantly define a font size for each element and the css won't be repeating itself.

We may find that we want to do more things like this such as margin and padding.  It's hard to determin what is best for the project right now, but after we test this stuff out, I will constantly try to find quicker and cleaner ways of presenting these sections.